chore: rename components folder to lowercase
